remove `mapbox-gl-js-test`
make all unit test to use `node:test` directly
move `assertEqualWithPrecision` to `function.test.js` since this is the
only place where it is needed
